Extract the code that prints warnings
